Modularizing flight tracking and GET endpoint
